requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
gradio>=4.0.0
modal>=1.2.0
